refactor(#1023): extract containsWakePhrase, add transcribeBlocking to VoiceInputController interface
