+++ core/voice/src/main/java/com/kernel/ai/core/voice/VoiceInputController.kt
@@ -22,4 +22,16 @@ interface VoiceInputController {
     suspend fun startListening(mode: VoiceCaptureMode): VoiceInputStartResult
 
     fun stopListening()
+
+    /**
+     * Transcribes [pcm] audio data synchronously and returns the transcript,
+     * or `null` when the engine does not support wake-word verification or
+     * transcription fails.
+     *
+     * Engines that need to perform async work (e.g. awaiting a mutex) can
+     * override this as `suspend` — callers in coroutine-unsafe contexts
+     * (e.g. [WakeWordService]'s verifyWindow) must wrap the call in
+     * [kotlinx.coroutines.runBlocking] when calling a [suspend] override.
+     */
+    suspend fun transcribeBlocking(pcm: ShortArray): String? = null
 }
