+++ core/voice/src/test/java/com/kernel/ai/core/voice/SherpaOnnxVoiceInputControllerTest.kt
@@ -3,7 +3,6 @@ package com.kernel.ai.core.voice
 import android.content.Context
 import android.media.AudioManager
 import kotlinx.coroutines.flow.MutableStateFlow
-import com.kernel.ai.core.voice.SherpaOnnxVoiceInputController.Companion.containsWakePhrase
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
@@ -78,47 +77,6 @@ class SherpaOnnxVoiceInputControllerTest {
         val result = controller.startListening(VoiceCaptureMode.Command)
 
         assertInstanceOf(VoiceInputStartResult.Unavailable::class.java, result)
-    }
-
-    // ── containsWakePhrase ─────────────────────────────────────────────────────
-
-    @Test
-    fun `containsWakePhrase matches Hey Jandal`() {
-        assertTrue("Hey Jandal".containsWakePhrase())
-        assertTrue("hey jandal".containsWakePhrase())
-        assertTrue("HEY JANDAL".containsWakePhrase())
-    }
-
-    @Test
-    fun `containsWakePhrase matches ASR variants`() {
-        assertTrue("a jandel".containsWakePhrase())
-        assertTrue("hey handel".containsWakePhrase())
-        assertTrue("Hey Handal".containsWakePhrase())
-        assertTrue("a hando".containsWakePhrase())
-    }
-
-    @Test
-    fun `containsWakePhrase rejects non-matches`() {
-        assertFalse("hello world".containsWakePhrase())
-        assertFalse("hey there".containsWakePhrase())
-        assertFalse("jandal".containsWakePhrase())
-        assertFalse("".containsWakePhrase())
-    }
-
-    @Test
-    fun `containsWakePhrase matches without whitespace`() {
-        // \s* allows zero whitespace between particles
-        assertTrue("heyjandal".containsWakePhrase())
-        assertTrue("heyjandel".containsWakePhrase())
-    }
-
-    @Test
-    fun `containsWakePhrase rejects embedded substrings`() {
-        // \b word boundaries prevent matching inside larger words
-        assertFalse("they jandal".containsWakePhrase())
-        assertFalse("hey jandalish".containsWakePhrase())
-        assertTrue("oh heyjandal who".containsWakePhrase())
-        assertTrue("say hey jandal now".containsWakePhrase())
     }
 
     // ── Helpers ───────────────────────────────────────────────────────────────
